move badge styles to css
